Fix for "Discount codes account limit and total limit not honored"
Closes issue 3783

In method cart/submitPayment any existing billing.paymentmethod (e.g.
paymentmethod for discount codes) was overwritten.

I guess this hasn't been working since introducing marketplace (just a
guess). The solution was to retain any already exsting
billing.paymentmethods in the array.

NOTICE: Because at various places (copyCartToOrder) the assumption is
that all relevant billing data (address, shopId, currency, etc.) is on
billing[0] , this does only work if it's appended at the end of the
array. I think it's not good to have this order predefined, but
hesitated to dig deeper here.

Also I'm not sure if the discout mechanism in general plays well with
the marketplace feature.