Fix for: Cloning products is not reactive
Resolves #3957

Changes:
Whenever a product is cloned an (unobserved) server side cursor in
publications/collections/products.js changes its results. But that does
not result in the changes being pushed to the client, because server
publications are not reactive. Pullling in one of several  meteor packages
available was not an option, therefore I've decided to trick the client
to resubscribe whenever that happens.